Added Q_tensor_shape to keep track of dimensions
Depending on whether the system is 2D, quasi-2D (i.e. 3D, but uniform in
        the z-direction), or 3D, several different quantities change.
In particular, the matrix dimension and number of degrees of freedom
change.
The system dimensionality is in an enum class NematicDimension, and the
specified shape quantities are static constexpr fields of structs which
are templated by each of the enum class's options.
I may add some mapping between the degrees of freedom and the matrix
entries, but I'm not sure yet.